Addressed Andre's feedback and fixed alt search results for resources
